requirements.txt
flasgger==0.9.7.1
Flask==2.3.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
kafka-python==2.0.2
pandas==2.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.1
